+++ packages/core/core/src/types.ts
@@ -26,8 +26,6 @@ export function Search(params: SearchParams, customQuery?: {}): any {
 export interface UseProduct<PRODUCT, PRODUCT_FILTERS, SORTING_OPTIONS> {
   products: ComputedProperty<PRODUCT[]>;
   totalProducts: ComputedProperty<number>;
-  availableFilters: ComputedProperty<PRODUCT_FILTERS>;
-  availableSortingOptions: ComputedProperty<SORTING_OPTIONS>;
   loading: ComputedProperty<boolean>;
   search: typeof Search;
   [x: string]: any;
@@ -181,7 +179,13 @@ export interface UseReview<REVIEW, REVIEWS_SEARCH_PARAMS, REVIEW_ADD_PARAMS> {
   addReview: (params: REVIEW_ADD_PARAMS) => Promise<void>;
   reviews: ComputedProperty<REVIEW>;
   loading: ComputedProperty<boolean>;
-  error: ComputedProperty<string | null>;
+  [x: string]: any;
+}
+
+export interface UseFacet<SEARCH_DATA> {
+  result: ComputedProperty<FacetSearchResult<SEARCH_DATA>>;
+  loading: ComputedProperty<boolean>;
+  search: (params?: AgnosticFacetSearchParams) => Promise<void>;
 }
 
 export interface ProductGetters<PRODUCT, PRODUCT_FILTER> {
@@ -277,6 +281,17 @@ export interface ReviewGetters<REVIEW, REVIEW_ITEM> {
   getReviewsPage: (review: REVIEW) => number;
 }
 
+export interface FacetsGetters<SEARCH_DATA, RESULTS, CRITERIA = any> {
+  getAll: (searchData: FacetSearchResult<SEARCH_DATA>, criteria?: CRITERIA) => AgnosticFacet[];
+  getGrouped: (searchData: FacetSearchResult<SEARCH_DATA>, criteria?: CRITERIA) => AgnosticGroupedFacet[];
+  getCategoryTree: (searchData: FacetSearchResult<SEARCH_DATA>) => AgnosticCategoryTree;
+  getSortOptions: (searchData: FacetSearchResult<SEARCH_DATA>) => AgnosticSort;
+  getProducts: (searchData: FacetSearchResult<SEARCH_DATA>) => RESULTS;
+  getPagination: (searchData: FacetSearchResult<SEARCH_DATA>) => AgnosticPagination;
+  getBreadcrumbs: (searchData: FacetSearchResult<SEARCH_DATA>) => AgnosticBreadcrumb[];
+  [getterName: string]: (element: any, options?: any) => unknown;
+}
+
 export interface AgnosticMediaGalleryItem {
   small: string;
   normal: string;
@@ -287,6 +302,7 @@ export interface AgnosticCategoryTree {
   label: string;
   slug?: string;
   items: AgnosticCategoryTree[];
+  isCurrent: boolean;
   [x: string]: unknown;
 }
 
@@ -360,4 +376,49 @@ export enum AgnosticOrderStatus {
   Complete = 'Complete',
   Cancelled = 'Cancelled',
   Refunded = 'Refunded'
+}
+
+export interface AgnosticFacet {
+  type: string;
+  id: string;
+  value: any;
+  attrName?: string;
+  count?: number;
+  selected?: boolean;
+  metadata?: any;
+}
+
+export interface AgnosticGroupedFacet {
+  id: string;
+  label: string;
+  count?: number;
+  options: AgnosticFacet[];
+}
+
+export interface AgnosticSort {
+  options: AgnosticFacet[];
+  selected: string;
+}
+
+export interface AgnosticPagination {
+  currentPage: number;
+  totalPages: number;
+  totalItems: number;
+  itemsPerPage: number;
+  pageOptions: number[];
+}
+
+export interface FacetSearchResult<S> {
+  data: S;
+  input: AgnosticFacetSearchParams;
+}
+
+export interface AgnosticFacetSearchParams {
+  categorySlug?: string;
+  page?: number;
+  itemsPerPage?: number;
+  sort?: string;
+  filters?: Record<string, string[]>;
+  metadata?: any;
+  [x: string]: any;
 }