+++ packages/core/nuxt-theme-module/theme/composables/index.ts
@@ -1,0 +1,5 @@
+import useUiHelpers from './useUiHelpers';
+
+export {
+  useUiHelpers
+};