+++ packages/core/docs/commercetools/changelog.md
@@ -6,10 +6,12 @@
 
 ### Changes
 
-* implemented useReviews on product page ([#4800](https://github.com/DivanteLtd/vue-storefront/issues/4800))
+- implemented useReviews on product page ([#4800](https://github.com/DivanteLtd/vue-storefront/issues/4800))
+- implemented faceting using useFacet factory (#4853)
 
 ## 0.1.0
 
 ### Changes
 
 * refactored setup using apiClientFactory ([#4777](https://github.com/DivanteLtd/vue-storefront/issues/4777))
+