+++ packages/core/docs/commercetools/changelog.md
@@ -2,6 +2,22 @@
 
 ## 0.2.0 (not released)
 
+## 0.1.1 (not released)
+
+### Breaking changes
+
+### Changes
+
+- #4616
+
+| Place | Before | 0.1.1 |
+|---|---|---|
+| cartGetters | - | implemented getCoupon getter |
+| useCart | Generic takes 4 arguments | Generic takes 3 arguments |
+| useCart | removeCoupon needs currentCart & coupon | removeCoupon needs only currentCart |
+
+# 0.1.0
+
 ### Breaking changes
 
 ### Changes